+++ src/ng-select/lib/value-utils.ts
@@ -1,3 +1,19 @@
+const unescapedHTMLExp = /[&<>"']/g;
+const hasUnescapedHTMLExp = RegExp(unescapedHTMLExp.source);
+const htmlEscapes = {
+    '&': '&amp;',
+    '<': '&lt;',
+    '>': '&gt;',
+    '"': '&quot;',
+    '\'': '&#39;'
+};
+
+export function escapeHTML(string: string) {
+    return (string && hasUnescapedHTMLExp.test(string)) ?
+        string.replace(unescapedHTMLExp, chr => htmlEscapes[chr]) :
+        string;
+}
+
 export function isDefined(value: any) {
     return value !== undefined && value !== null;
 }